Add user relation in actions and rest endpoint on CRUD Todos

+++ src/app/api/todos/route.ts
@@ -1,3 +1,4 @@
+import { getServerUserSession } from '@/auth/actions/auth-actions'
 import prisma from '@/lib/prisma'
 import { NextResponse } from 'next/server'
 import * as yup from 'yup'
@@ -27,9 +28,15 @@ const postSchema = yup.object({
 export async function POST(request: Request) {
   const allowedFields = Object.keys(postSchema.describe().fields).join(', ')
 
+  const user = await getServerUserSession()
+
+  if (!user) {
+    return NextResponse.json({ message: 'Unauthorized' }, { status: 401 })
+  }
+
   try {
-    const body = await postSchema.validate(await request.json())
-    const todo = await prisma.todo.create({ data: body })
+    const { complete, description } = await postSchema.validate(await request.json())
+    const todo = await prisma.todo.create({ data: { description, complete, userId: user!.id } })
 
     return NextResponse.json(todo)
   } catch (error) {
@@ -42,9 +49,15 @@ export async function POST(request: Request) {
 
 // Elimina todos los Todos en estado Complete = true
 export async function DELETE(request: Request) {
+  const user = await getServerUserSession()
+
+  if (!user) {
+    return NextResponse.json({ message: 'Unauthorized' }, { status: 401 })
+  }
+
   try {
     await prisma.todo.deleteMany({
-      where: { complete: true }
+      where: { complete: true, userId: user!.id }
     })
 
     return NextResponse.json({ message: 'Todos with complete = true deleted' })
